+++ src/Worker.cc
@@ -58,7 +58,7 @@ void Worker::allocate_arrays() {
   cycleused = new bool[graph.numstates + 1];
   deadstates = new int[graph.numstates + 1];
 
-  for (int i = 0; i <= graph.numstates; ++i) {
+  for (size_t i = 0; i <= graph.numstates; ++i) {
     pattern[i] = -1;
     used[i] = 0;
     cycleused[i] = false;
@@ -269,9 +269,8 @@ void Worker::load_work_assignment(const WorkAssignment& wa) {
   assert(root_throwval_options.size() > 0);
   assert(pos == 0);
 
-  for (int i = 0; i <= graph.numstates; ++i) {
-    pattern[i] = (i < static_cast<int>(wa.partial_pattern.size())) ?
-        wa.partial_pattern[i] : -1;
+  for (size_t i = 0; i <= graph.numstates; ++i) {
+    pattern[i] = (i < wa.partial_pattern.size()) ? wa.partial_pattern[i] : -1;
   }
 }
 
@@ -285,7 +284,7 @@ WorkAssignment Worker::get_work_assignment() const {
   wa.end_state = end_state;
   wa.root_pos = root_pos;
   wa.root_throwval_options = root_throwval_options;
-  for (int i = 0; i <= graph.numstates; ++i) {
+  for (size_t i = 0; i <= graph.numstates; ++i) {
     if (pattern[i] == -1)
       break;
     wa.partial_pattern.push_back(pattern[i]);
@@ -372,7 +371,7 @@ WorkAssignment Worker::split_work_assignment_takefraction(double f,
   wa.start_state = start_state;
   wa.end_state = start_state;
   wa.root_pos = root_pos;
-  for (int i = 0; i < root_pos; ++i)
+  for (size_t i = 0; i < root_pos; ++i)
     wa.partial_pattern.push_back(pattern[i]);
 
   // ensure the throw value at `root_pos` isn't on the list of throw options
@@ -429,7 +428,7 @@ WorkAssignment Worker::split_work_assignment_takefraction(double f,
 
     // have to scan from the beginning because we don't record the traversed
     // states as we build the pattern
-    for (int pos2 = 0; pos2 <= pos; ++pos2) {
+    for (size_t pos2 = 0; pos2 <= pos; ++pos2) {
       const int throwval = pattern[pos2];
       for (col = 0; col < graph.outdegree[from_state]; ++col) {
         if (throwval == graph.outthrowval[from_state][col])
@@ -484,7 +483,7 @@ void Worker::gen_patterns() {
     blocklength = 0;
     max_possible = maxlength;
     exitcyclesleft = 0;
-    for (int i = 0; i <= graph.numstates; ++i) {
+    for (size_t i = 0; i <= graph.numstates; ++i) {
       used[i] = 0;
       cycleused[i] = false;
       deadstates[i] = 0;
@@ -529,7 +528,7 @@ void Worker::gen_patterns() {
     if (config.mode != RunMode::SUPER_SEARCH) {
       // verify that we didn't prune the shift throw at column 0, which is
       // assumed in some of the gen_loops_xxx() methods below
-      for (int i = 1; i <= graph.numstates; ++i) {
+      for (size_t i = 1; i <= graph.numstates; ++i) {
         assert(graph.outthrowval[i][0] == 0 ||
             graph.outthrowval[i][0] == graph.h);
       }
